+++ src/libcsg/topology.cc
@@ -138,8 +138,8 @@ void Topology::Add(Topology *top)
     
     for(bead=top->_beads.begin(); bead!=top->_beads.end(); ++bead) {
         Bead *bi = *bead;
-        BeadType *type =  GetOrCreateBeadType(bi->getType()->getName());
-        CreateBead(bi->getSymmetry(), bi->getName(), type, bi->getResnr()+res0, bi->getMass(), bi->getQ());
+        weak_ptr<BeadType> weak_type = bi->getType();
+        CreateBead(bi->getSymmetry(), bi->getName(), weak_type, bi->getResnr()+res0, bi->getMass(), bi->getQ());
     }
     
     for(res=top->_residues.begin();res!=top->_residues.end(); ++res) {
@@ -176,8 +176,8 @@ void Topology::CopyTopologyData(Topology *top)
     // create all beads
     for(it_bead=top->_beads.begin(); it_bead!=top->_beads.end(); ++it_bead) {
         Bead *bi = *it_bead;
-        BeadType *type =  GetOrCreateBeadType(bi->getType()->getName());
-        Bead *bn = CreateBead(bi->getSymmetry(), bi->getName(), type, bi->getResnr(), bi->getMass(), bi->getQ());
+        weak_ptr<BeadType> weak_type = bi->getType();
+        Bead *bn = CreateBead(bi->getSymmetry(), bi->getName(), weak_type, bi->getResnr(), bi->getMass(), bi->getQ());
         bn->setOptions(bi->Options());
     }
 
@@ -214,9 +214,13 @@ void Topology::RenameBeadType(string name, string newname)
 {
     BeadContainer::iterator bead;
     for(bead=_beads.begin(); bead!=_beads.end(); ++bead) {
-      BeadType *type =  GetOrCreateBeadType((*bead)->getType()->getName());
-      if (wildcmp(name.c_str(),(*bead)->getType()->getName().c_str())) {
-	type->setName(newname);
+      weak_ptr<BeadType> weak_type =  (*bead)->getType();
+      if(shared_ptr<BeadType> shared_type = weak_type.lock()){
+        if (wildcmp(name.c_str(),shared_type->getName().c_str())) {
+          shared_type->setName(newname);
+        }
+      }else{
+        assert(!"bead type is not accessible while attempting to rename.");
       }
     }
 }
@@ -225,8 +229,13 @@ void Topology::SetBeadTypeMass(string name, double value)
 {
     BeadContainer::iterator bead;
     for(bead=_beads.begin(); bead!=_beads.end(); ++bead) {
-      if (wildcmp(name.c_str(),(*bead)->getType()->getName().c_str())) {
-	(*bead)->setMass(value);
+      weak_ptr<BeadType> weak_type = (*bead)->getType();
+      if(shared_ptr<BeadType> type = weak_type.lock()){
+        if (wildcmp(name.c_str(),type->getName().c_str())) {
+          (*bead)->setMass(value);
+        }
+      }else{
+        assert(!"bead type no longer exist memory error.");
       }
     }
 
@@ -274,21 +283,19 @@ std::list<Interaction *> Topology::InteractionsInGroup(const string &group)
 }
 
 
-BeadType *Topology::GetOrCreateBeadType(string name)
+weak_ptr<BeadType> Topology::GetOrCreateBeadType(string name)
 {
     map<string, int>::iterator iter;
     
     iter = _beadtype_map.find(name);
     if(iter == _beadtype_map.end()) {
-        BeadType *bt = new BeadType(this, _beadtypes.size(), name);
-        _beadtypes.push_back(bt);
-        _beadtype_map[name] = bt->getId();
-        return bt;
+        _beadtypes.push_back(shared_ptr<BeadType>( new BeadType(this,_beadtypes.size(),name)));
+        _beadtype_map[name] = _beadtypes.back()->getId();
+        return (_beadtypes.back());
     }
-    else {
-        return _beadtypes[(*iter).second];
-    }
-    return NULL;
+
+    return  (_beadtypes[(*iter).second]);
+    
 }
 
 vec Topology::BCShortestConnection(const vec &r_i, const vec &r_j) const